refactor: rextract DataFetcher
