+++ src/modes/correlation/Correlation.svelte
@@ -12,7 +12,7 @@
     currentLag,
     metaDataManager,
   } from '../../stores';
-  import { DataFetcher, DateParam, RegionParam, SensorParam } from '../../stores/params';
+  import { DateParam, RegionParam, SensorParam } from '../../stores/params';
   import RegionDatePicker from '../../components/RegionDatePicker.svelte';
   import IndicatorPicker from '../../components/IndicatorPicker.svelte';
   import IndicatorCorrelationChart from './IndicatorCorrelationChart.svelte';
@@ -23,6 +23,7 @@
   import AboutSection from '../../components/AboutSection.svelte';
   import { fetchSingleCorrelations } from '../../data/correlation';
   import throttle from 'lodash-es/throttle';
+  import { DataFetcher } from '../../stores/DataFetcher';
 
   $: primary = new SensorParam($currentSensorEntry, $metaDataManager);
   $: secondary = new SensorParam($currentSensorEntry2, $metaDataManager, currentSensor2);
