+++ src/modes/correlation/IndicatorCorrelationChart.svelte
@@ -33,7 +33,7 @@
   export let region;
 
   /**
-   * @type {import("../../stores/params").DataFetcher}
+   * @type {import("../../stores/DataFetcher").DataFetcher}
    */
   export let fetcher;
