+++ src/modes/dashboard/widgets/DateParallelCoordinatesWidget.svelte
@@ -29,7 +29,7 @@
   export let highlight = null;
 
   /**
-   * @type {import("../../../stores/params").DataFetcher}
+   * @type {import("../../../stores/DataFetcher").DataFetcher}
    */
   const fetcher = getContext('fetcher');
