+++ src/modes/dashboard/widgets/KPITrendWidget.svelte
@@ -38,7 +38,7 @@
   export let initialState = DEFAULT_STATE;
 
   /**
-   * @type {import("../../../stores/params").DataFetcher}
+   * @type {import("../../../stores/DataFetcher").DataFetcher}
    */
   const fetcher = getContext('fetcher');
