+++ src/modes/dashboard/widgets/KPIWidget.svelte
@@ -31,7 +31,7 @@
   };
 
   /**
-   * @type {import("../../../stores/params").DataFetcher}
+   * @type {import("../../../stores/DataFetcher").DataFetcher}
    */
   const fetcher = getContext('fetcher');
