+++ src/modes/dashboard/widgets/LineChartWidget.svelte
@@ -70,7 +70,7 @@
   }
 
   /**
-   * @type {import("../../../stores/params").DataFetcher}
+   * @type {import("../../../stores/DataFetcher").DataFetcher}
    */
   const fetcher = getContext('fetcher');
