+++ src/modes/dashboard/widgets/MapChartWidget.svelte
@@ -46,7 +46,7 @@
   };
 
   /**
-   * @type {import("../../../stores/params").DataFetcher}
+   * @type {import("../../../stores/DataFetcher").DataFetcher}
    */
   const fetcher = getContext('fetcher');
