+++ src/modes/dashboard/widgets/RegionTableWidget.svelte
@@ -28,7 +28,7 @@
   export let initialState = DEFAULT_STATE;
 
   /**
-   * @type {import("../../../stores/params").DataFetcher}
+   * @type {import("../../../stores/DataFetcher").DataFetcher}
    */
   const fetcher = getContext('fetcher');
