+++ src/modes/dashboard/widgets/Template.svelte
@@ -21,7 +21,7 @@
   export let highlight = null;
 
   /**
-   * @type {import("../../../stores/params").DataFetcher}
+   * @type {import("../../../stores/DataFetcher").DataFetcher}
    */
   const fetcher = getContext('fetcher');
