+++ src/modes/summary/AllIndicatorOverview.svelte
@@ -13,7 +13,7 @@
    */
   export let region;
   /**
-   * @type {import("../../stores/params").DataFetcher}
+   * @type {import("../../stores/DataFetcher").DataFetcher}
    */
   export let fetcher;
