+++ src/modes/survey-results/IndicatorRevision.svelte
@@ -19,7 +19,7 @@
    */
   export let region;
   /**
-   * @type {import("../../stores/params").DataFetcher}
+   * @type {import("../../stores/DataFetcher").DataFetcher}
    */
   export let fetcher;
