+++ src/modes/survey-results/IndicatorRevisions.svelte
@@ -15,7 +15,7 @@
    */
   export let region;
   /**
-   * @type {import("../../stores/params").DataFetcher}
+   * @type {import("../../stores/DataFetcher").DataFetcher}
    */
   export let fetcher;
 </script>
