+++ src/modes/survey-results/SurveyQuestion.svelte
@@ -27,7 +27,7 @@
    */
   export let region;
   /**
-   * @type {import("../../stores/params").DataFetcher}
+   * @type {import("../../stores/DataFetcher").DataFetcher}
    */
   export let fetcher;
